Split stdlib filter tests into modules
Break the oversized std_filter_tests integration suite into smaller modules so each file stays under 400 lines. Update Cargo features to use caret-style defaults without explicit hats while keeping hash coverage intact.
